transcription-service/requirements.txt
flask
torch
openai-whisper